Remove all DropInFlowManagingMock duplicates
