+++ Tests/IntegrationTests/DropIn Tests/PaymentMethodList/PaymentMethodListViewModelTests.swift
@@ -281,85 +281,6 @@ struct PaymentMethodListViewModelTests {
         }
     }
 
-    // MARK: - Mocks
-
-    private class DropInFlowManagingMock: DropInFlowManaging {
-        var submitFromActionPresenterCallsCount = 0
-        var submitFromActionPresenterReceivedArguments: (
-            data: PaymentComponentData,
-            component: PaymentComponent,
-            actionPresenter: ActionPresenter
-        )?
-
-        func submit(
-            _ data: PaymentComponentData,
-            from component: PaymentComponent,
-            actionPresenter: ActionPresenter
-        ) {
-            submitFromActionPresenterCallsCount += 1
-            submitFromActionPresenterReceivedArguments = (data, component, actionPresenter)
-        }
-
-        var failWithFromCallsCount = 0
-        var failWithFromReceivedArguments: (error: Error, component: PaymentComponent)?
-
-        func fail(with error: Error, from component: PaymentComponent) {
-            failWithFromCallsCount += 1
-            failWithFromReceivedArguments = (error, component)
-        }
-
-        var cancelComponentCallsCount = 0
-        var cancelComponentReceivedComponent: PaymentComponent?
-
-        func cancel(component: PaymentComponent) {
-            cancelComponentCallsCount += 1
-            cancelComponentReceivedComponent = component
-        }
-
-        var handleActionCallsCount = 0
-        var handleActionReceivedAction: Action?
-
-        func handle(action: Action) {
-            handleActionCallsCount += 1
-            handleActionReceivedAction = action
-        }
-    }
-
-    private class PaymentMethodListRoutingMock: PaymentMethodListRouting {
-        var presentComponentCallsCount = 0
-        var presentComponentReceivedComponent: PaymentComponent?
-
-        func present(component: PaymentComponent) {
-            presentComponentCallsCount += 1
-            presentComponentReceivedComponent = component
-        }
-
-        var presentViewControllerCallsCount = 0
-        var presentViewControllerReceivedViewController: UIViewController?
-
-        func present(viewController: UIViewController) {
-            presentViewControllerCallsCount += 1
-            presentViewControllerReceivedViewController = viewController
-        }
-
-        var presentActionComponentOnCancelCallsCount = 0
-        var presentActionComponentOnCancelReceivedArguments: (actionComponent: PresentableComponent, onCancel: (() -> Void)?)?
-
-        func present(actionComponent: PresentableComponent, onCancel: (() -> Void)?) {
-            presentActionComponentOnCancelCallsCount += 1
-            presentActionComponentOnCancelReceivedArguments = (actionComponent, onCancel)
-        }
-
-        var dismissCompletionCallsCount = 0
-        var dismissCompletionReceivedCompletion: (() -> Void)?
-
-        func dismiss(completion: (() -> Void)?) {
-            dismissCompletionCallsCount += 1
-            dismissCompletionReceivedCompletion = completion
-            completion?()
-        }
-    }
-
     // MARK: - Helpers
 
     private func makeSUT(includeApplePay: Bool = true) -> (
